AISnakeGame/requirement.txt
argon2-cffi==20.1.0
autopep8==1.5.7
ipykernel==5.5.4
matplotlib==3.4.2
nbconvert==6.0.7
openpyxl==3.0.7
pandas==1.2.4
pipdeptree==2.0.0
prometheus-client==0.10.1
pygame==2.0.1
pylint==2.8.2
Send2Trash==1.5.0
terminado==0.9.4
torchvision==0.9.1
